@@ -1,11 +1,13 @@
 {
   "name": "open-synthesis",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Open platform for CIA-style intelligence analysis",
   "private": true,
   "main": "index.js",
   "scripts": {
-    "postinstall": "webpack --config ./webpack.config.js",
+    "build": "npx webpack --config webpack.config.prod.js",
+    "watch": "NODE_ENV='development' npx webpack --config webpack.config.dev.js --watch",
+    "postinstall": "npm run build",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -24,18 +26,23 @@
   },
   "homepage": "https://github.com/twschiller/open-synthesis#readme",
   "dependencies": {
-    "bootstrap": "^3.3.7",
+    "@babel/core": "^7.7.7",
+    "@babel/preset-env": "^7.7.7",
+    "babel-loader": "^8.0.6",
+    "bootstrap": "^3.4.1",
     "bootstrap-datepicker": "^1.6.4",
-    "css-loader": "^0.25.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
+    "css-loader": "^3.4.1",
+    "eslint": "^6.8.0",
+    "file-loader": "^5.0.2",
     "jquery": "^3.4.1",
-    "selectize": "^0.12.3",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.2",
-    "webpack-bundle-tracker": "0.0.93"
-  },
-  "devDependencies": {
-    "eslint": "^3.5.0"
+    "mini-css-extract-plugin": "^0.8.2",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "selectize": "^0.12.6",
+    "style-loader": "^1.1.2",
+    "terser-webpack-plugin": "^2.3.1",
+    "webpack": "^4.41.5",
+    "webpack-bundle-tracker": "^0.4.3",
+    "webpack-cli": "^3.3.10",
+    "webpack-merge": "^4.2.2"
   }
 }
