@@ -259,7 +259,7 @@ STATIC_URL = '/static/'
 # Extra places for collectstatic to find static files.
 STATICFILES_DIRS = (
     os.path.join(PROJECT_ROOT, 'static'),
-    os.path.join(BASE_DIR, 'assets')
+    os.path.join(BASE_DIR, 'openach-frontend', 'bundles')
 )
 
 # Simplified static file serving.
@@ -269,7 +269,8 @@ STATICFILES_STORAGE = 'whitenoise.storage.CompressedManifestStaticFilesStorage'
 # https://github.com/owais/django-webpack-loader#default-configuration
 WEBPACK_LOADER = {
     'DEFAULT': {
-        'BUNDLE_DIR_NAME': 'bundles/',
+        "CACHE": not DEBUG,
+        'BUNDLE_DIR_NAME': '/',
         'STATS_FILE': os.path.join(BASE_DIR, 'webpack-stats.json'),
     }
 }
